fix(nfs): Fix a role failure when a configured nfs share is removed
In that case the nfs-kernel-server.service failed to restart,
aborting the execution of the whole role.
It fails because for one of the exports configured in
/etc/exports.d/zfs-storage-ansible.exports
the directory does not exist.
The zfs-storage-ansible.exports is never updated and the old export removed
because the role never gets to that point.

Now the role only restarts the service at the end,
after the zfs-storage-ansible.exports is updated